# openapi.yaml
openapi: 3.1.0
info:
  title: PORTA'M Server API
  summary: Public transport validation system backoffice for all of its modules.
  description: >
    This document captures the current Express routes mounted under `/api` and the helper
    endpoints defined in `server.js`. Each controller in `controllers/` documents the underlying
    Supabase tables that back the module. All responses wrap the returned data in a `success` flag and either
    a resource payload or a textual `error` field, except for the validation workflow which returns
    localized status messages, as this information is shown to the user as a feedback while validating.
  version: 1.0.0
servers:
  - url: https://portam-server.vercel.app/
    description: Production server
...
paths:
...
  /api/users/{id}:
    parameters:
      - $ref: '#/components/parameters/UserId'
    get:
      tags: [Users]
      operationId: getUser
      summary: Get user by ID
      responses:
        '200':
          description: User detail (password removed)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '500':
          $ref: '#/components/responses/ServerError'
    put:
      tags: [Users]
      operationId: updateUser
      summary: Update user profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserUpdateRequest'
      responses:
        '200':
          description: User updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserWriteResponse'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '500':
          $ref: '#/components/responses/ServerError'
    delete:
      tags: [Users]
      operationId: deleteUser
      summary: Delete user
      description: >
        Removes the user and cleans all memberships in `user_groups`.
      responses:
        '200':
          description: User removed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserDeleteResponse'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '500':
          $ref: '#/components/responses/ServerError'
  /api/users/create:
    post:
      tags: [Users]
      operationId: createUser
      summary: Create user
      description: >
        Generates a random 12-digit `id`, inserts the user, and optionally assigns them to groups
        (creating `user_groups` records and respecting group expiration rules).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserCreateRequest'
      responses:
        '201':
          description: User created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserWriteResponse'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '500':
          $ref: '#/components/responses/ServerError'
  /api/users/{id}/groups:
    parameters:
      - $ref: '#/components/parameters/UserId'
    post:
      tags: [UserGroups]
      operationId: addGroupToUser
      summary: Add group to user
      description: >
        Assigns a group to a user, calculating expiration dates based on birthdate-aware special groups
        (16, 30, 65) or the group's `expiration` days unless a manual expiration is provided.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserGroupAssignmentRequest'
      responses:
        '201':
          description: Membership created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserGroupAssignmentResponse'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '500':
          $ref: '#/components/responses/ServerError'
    get:
      tags: [UserGroups]
      operationId: listUserGroups
      summary: List user groups
      responses:
        '200':
          description: Groups attached to the user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserGroupsResponse'
        '500':
          $ref: '#/components/responses/ServerError'
  /api/users/{id}/groups/{groupId}:
    parameters:
      - $ref: '#/components/parameters/UserId'
      - $ref: '#/components/parameters/GroupId'
    delete:
      tags: [UserGroups]
      operationId: removeGroupFromUser
      summary: Remove group from user
      responses:
        '200':
          description: Membership removed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserGroupRemovalResponse'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '500':
          $ref: '#/components/responses/ServerError'
  /api/users/{id}/suports:
    parameters:
      - $ref: '#/components/parameters/UserId'
    get:
      tags: [Suports]
      operationId: listUserSuports
      summary: List user suports
      responses:
        '200':
          description: Suports attached to the user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuportListResponse'
        '500':
          $ref: '#/components/responses/ServerError'
    post:
      tags: [Suports]
      operationId: addSuportToUser
      summary: Attach suport to user
      description: >
        Creates a suport if the UID does not exist or re-assigns an unbound suport to the user,
        stamping the activation time.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SuportCreateRequest'
      responses:
        '201':
          description: Suport created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuportResponse'
        '200':
          description: Suport reassigned
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuportResponse'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '500':
          $ref: '#/components/responses/ServerError'
  /api/users/{id}/suports/{uid}:
    parameters:
      - $ref: '#/components/parameters/UserId'
      - name: uid
        in: path
        required: true
        description: Unique suport identifier (varchar UID).
        schema:
          type: string
    get:
      tags: [Suports]
      operationId: getUserSuport
      summary: Get suport details
      responses:
        '200':
          description: Suport record
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuportResponse'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '500':
          $ref: '#/components/responses/ServerError'
    delete:
      tags: [Suports]
      operationId: removeSuportFromUser
      summary: Remove suport from user
      responses:
        '200':
          description: Suport detached
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuportRemovalResponse'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '500':
          $ref: '#/components/responses/ServerError'
...
components:
  parameters:
    UserId:
      name: id
      in: path
      required: true
      description: 12-digit user identifier.
      schema:
        type: string
        example: '123456789012'
    GroupId:
      name: groupId
      in: path
      required: true
      description: Group identifier.
      schema:
        type: string
        example: '16'
...
  responses:
    BadRequestError:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFoundError:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    ServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: string
          example: Something went wrong
      required: [success, error]
    User:
      type: object
      properties:
        id:
          type: string
          description: Random 12-digit identifier.
          example: '123456789012'
        name:
          type: string
        surname:
          type: string
        gov_id:
          type: string
          description: Government-issued identifier.
        email:
          type: string
          format: email
        phone:
          type: string
        birthdate:
          type: string
          format: date
      required: [id, name, surname, gov_id, email, phone, birthdate]
...
    UserResponse:
      type: object
      properties:
        success:
          type: boolean
        user:
          $ref: '#/components/schemas/User'
      required: [success, user]
    UserCreateRequest:
      type: object
      properties:
        name:
          type: string
        surname:
          type: string
        gov_id:
          type: string
        email:
          type: string
          format: email
        phone:
          type: string
        birthdate:
          type: string
          format: date
        groups:
          type: array
          description: Optional list of group IDs to insert into `user_groups`.
          items:
            type: string
      required: [name, surname, gov_id, email, phone, birthdate]
    UserUpdateRequest:
      type: object
      properties:
        name:
          type: string
        surname:
          type: string
        gov_id:
          type: string
        email:
          type: string
          format: email
        phone:
          type: string
        birthdate:
          type: string
          format: date
    UserWriteResponse:
      type: object
      properties:
        success:
          type: boolean
        user_id:
          type: string
          example: '123456789012'
      required: [success, user_id]
    UserDeleteResponse:
      allOf:
        - $ref: '#/components/schemas/UserWriteResponse'
    UserGroupMembership:
      type: object
      properties:
        user:
          type: string
        group:
          type: string
        expiration:
          type: string
          format: date
          nullable: true
      required: [user, group]
    UserGroupAssignmentRequest:
      type: object
      properties:
        groupId:
          type: string
          description: Target group identifier.
        manualExpiration:
          type: string
          format: date
          nullable: true
          description: Optional custom expiration date (YYYY-MM-DD).
      required: [groupId]
    UserGroupAssignmentResponse:
      type: object
      properties:
        success:
          type: boolean
        user_id:
          type: string
        group_id:
          type: string
        expiration:
          type: string
          format: date
          nullable: true
      required: [success, user_id, group_id]
    UserGroupRemovalResponse:
      type: object
      properties:
        success:
          type: boolean
        user_id:
          type: string
        group_id:
          type: string
      required: [success, user_id, group_id]
    UserGroupsResponse:
      type: object
      properties:
        success:
          type: boolean
        user_id:
          type: string
        groups:
          type: array
          items:
            $ref: '#/components/schemas/UserGroupMembership'
      required: [success, user_id, groups]
    Suport:
      type: object
      properties:
        uid:
          type: string
        user:
          type: string
          nullable: true
        activation:
          type: string
          format: date-time
          nullable: true
      required: [uid]
    SuportListResponse:
      type: object
      properties:
        success:
          type: boolean
        user_id:
          type: string
        suports:
          type: array
          items:
            $ref: '#/components/schemas/Suport'
      required: [success, user_id, suports]
    SuportCreateRequest:
      type: object
      properties:
        uid:
          type: string
          description: Unique suport identifier to attach.
      required: [uid]
    SuportResponse:
      type: object
      properties:
        success:
          type: boolean
        suport:
          $ref: '#/components/schemas/Suport'
      required: [success, suport]
    SuportRemovalResponse:
      type: object
      properties:
        success:
          type: boolean
        uid:
          type: string
      required: [success, uid]
...